Obsolete but still tolerated SEP DD (and UNIT) parameter added.
Ran across the SEP parameter of the DD statement and subparameter of the UNIT DD parameter whilst testing with JCL taken from SYS1.PROCLIB in the Hercules emulator.  Thankfully some folks (including the much-and-still-missed Phil Payne) were discussing SEP back in 1999 on the ibm-main listserv so I could understand why it wasn't in the current IBM documentation.  Its former purpose is documented in the lexer.