javadoc and bug fixes
EmbeddedEquality incorporated into SingleOrMultipleValue parameterStrings, determined ParameterString cannot be replaced with KeywordValueWrapper, fixed bug in DdStatement where toCSV did not correctly identify DD * and the catch-all did not correctly identify itself.  Fixed JCLPPLexer to emit EXEC statement parameters to prevent all symbolics from being treated as part of that being executed.  Fixed KeywordOrSymbolicWrapper instantiation to, in the event of a null parameter of the form "DATACLAS=" or similar, to log a warning with a stack trace and not attempt to continue with initialization.