testing is now more robust, some Makefile doc
This was an adventure.  It turns out the native Java logging spits out its
messages to stderr, as does ANTLR with its messages.  In order to separate
these, I began by changing the Java logging to put its messages to stdout.
This led to the discovery that doing so seemed to result in their being
written in 8K blocks which was definitely not what I wanted.  So I undid that
and created StdoutLexerErrorListener and StdoutParserErrorListener so the
ANTLR messages would go to stdout.  All this was so I could, with some
confidence, determine there had been no ANTLR messages indicating a problem
with lexing or parsing.  And in the process I did a make clean which got
rid of an impressive amount of cruft laying about in the ./class directory,
left over from various iterations of the lexer and parser and going back who
knows how long.