a solution(?) to the SYSPROC.DSNUTILx third argument problem
This potential solution to the problem of the third argument
to the SYSPROC.DSNUTILx stored procedures requires somewhat
extensive and lower level action code in the DSNUTIL_APOS
and DSNUTIL_QUOTE tokens.

I'm not happy with this, but it does seem to be the only way
I've tried to date that works in all cases and doesn't break
anything else.